requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.5
django-admin==2.0.1
django-cpf==0.1.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.12.4
idna==3.2
install==1.3.4
model-mommy==2.0.0
psycopg2==2.9.1
pytz==2021.1
requests==2.26.0
screen==1.0.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
xlwt==1.3.0
